+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [2019.0928]
+
+  - Setup coverage testing
+  - Runtime camera permission check, refs #5
+  - Publish `xcamera` demo binary, refs #9
+  - Add Python3.7 support, refs #13
+
 ## [2019.0911]
 
   - Move to new garden layout